requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
async-timeout==3.0.1
attrs==22.2.0
boto==2.49.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.7
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.13.4
frozenlist==1.3.3
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.22.2
idna==3.4
multidict==6.0.4
orjson==3.10.3
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.64.1
typing_extensions==4.11.0
urllib3==1.26.14
yarl==1.8.2
